+++ changelog.md
@@ -131,6 +131,7 @@ Release Date: 11/9/2017
 
     Views
         - views/_menubar.php: added dropdown menu to navbar for selecting roles, displays role if set (TE)
+        - Made tooltip look pretty for each flight number
 
     Config
         - app/config/autoload.php, config/config.php: enabled sessions (TE)